+++ src/main/java/woojooin/planit/global/config/RootConfig.java
@@ -30,7 +30,7 @@ import javax.sql.DataSource;
     })
 @MapperScan(basePackages  = {
     "woojooin.planit.domain.member.mapper",
-    "woojooin.planit.domain.account.isa.mapper"  // ISA 계좌 mapper 추가
+    "woojooin.planit.domain.object.isa.mapper"  // ISA 계좌 mapper 추가
 })
 @Slf4j
 @EnableTransactionManagement
